refactor: Orders moved inside history package for orders.
