+++ app/src/main/java/com/coffee/shop/presentation/orders/history/OrdersUIEvents.kt
@@ -1,4 +1,4 @@
-package com.coffee.shop.presentation.orders
+package com.coffee.shop.presentation.orders.history
 
 sealed interface OrdersUIEvents {
     data object OnReorder : OrdersUIEvents
